# apps/api/src/swagger/swagger_ui.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "POP Api",
		"description": "Documentation sur les webservices de l'application Api"
	},
	"host": "localhost:3000",
	"basePath": "/",
...
	"schemes": ["http"],
	"securityDefinitions": {
		"apiKeyAuth": {
			"type": "apiKey",
			"in": "header",
			"name": "X-API-KEY",
			"description": "any description..."
		}
	},
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/joconde/{ref}": {
			"put": {
				"tags": ["Joconde"],
				"description": "Modification de la notice Joconde",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"notice": {
									"example": "any"
								},
								"updateMode": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"get": {
				"tags": ["Joconde"],
				"description": "Retourne les informations de la notice Joconde",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "Référence de la notice Joconde"
					}
				],
				"responses": {
					"200": {
						"description": "Récupération des informations avec succés",
						"schema": {
							"$ref": "#/definitions/GetJoconde"
						}
					},
					"404": {
						"description": "Document non trouvé",
						"schema": {
							"type": "object",
							"properties": {
								"success": {
									"type": "boolean",
									"example": false
								},
								"msg": {
									"type": "string",
									"example": "Document introuvable"
								}
							},
							"xml": {
								"name": "main"
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Joconde"],
				"description": "Suppression de la notice Joconde",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/joconde": {
			"post": {
				"tags": ["Joconde"],
				"description": "Création de la notice Joconde",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"notice": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"GetJoconde": {
			"type": "object",
			"properties": {
				"POP_COORDONNEES": {
					"type": "object",
					"properties": {
						"lat": {
							"type": "number",
							"example": 44.931125
						},
						"lon": {
							"type": "number",
							"example": 4.889563
						}
					}
				},
				"PRODUCTEUR": {
					"type": "string",
					"example": ""
				},
				"BASE": {
					"type": "string",
					"example": ""
				},
				"CONTIENT_IMAGE": {
					"type": "string",
					"example": "oui"
				},
				"POP_CONTIENT_GEOLOCALISATION": {
					"type": "string",
					"example": "oui"
				},
				"POP_FLAGS": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"POP_IMPORT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REFMIS": {
					"type": "string",
					"example": ""
				},
				"MANQUANT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"MANQUANT_COM": {
					"type": "string",
					"example": ""
				},
				"ADPT": {
					"type": "string",
					"example": ""
				},
				"APPL": {
					"type": "string",
					"example": ""
				},
				"APTN": {
					"type": "string",
					"example": ""
				},
				"ATTR": {
					"type": "string",
					"example": ""
				},
				"AUTR": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"BIBL": {
					"type": "string",
					"example": ""
				},
				"COMM": {
					"type": "string",
					"example": ""
				},
				"CONTACT": {
					"type": "string",
					"example": ""
				},
				"COOR": {
					"type": "string",
					"example": ""
				},
				"COPY": {
					"type": "string",
					"example": ""
				},
				"DACQ": {
					"type": "string",
					"example": ""
				},
				"DATA": {
					"type": "string",
					"example": ""
				},
				"DATION": {
					"type": "string",
					"example": ""
				},
				"DDPT": {
					"type": "string",
					"example": ""
				},
				"DECV": {
					"type": "string",
					"example": ""
				},
				"DENO": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"DEPO": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"DESC": {
					"type": "string",
					"example": ""
				},
				"DESY": {
					"type": "string",
					"example": ""
				},
				"DIFFU": {
					"type": "string",
					"example": ""
				},
				"DIMS": {
					"type": "string",
					"example": ""
				},
				"DMAJ": {
					"type": "string",
					"example": ""
				},
				"DMIS": {
					"type": "string",
					"example": ""
				},
				"DOMN": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"DREP": {
					"type": "string",
					"example": ""
				},
				"ECOL": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"EPOQ": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"ETAT": {
					"type": "string",
					"example": ""
				},
				"EXPO": {
					"type": "string",
					"example": ""
				},
				"GENE": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"GEOHI": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"HIST": {
					"type": "string",
					"example": ""
				},
				"IMAGE": {
					"type": "string",
					"example": "oui"
				},
				"IMG": {
					"type": "array",
					"example": ["", ""],
					"items": {
						"type": "string"
					}
				},
				"INSC": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"INV": {
					"type": "string",
					"example": ""
				},
				"LABEL": {
					"type": "string",
					"example": ""
				},
				"LABO": {
					"type": "string",
					"example": ""
				},
				"LARC": {
					"type": "string",
					"example": ""
				},
				"LIEUX": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REGION": {
					"type": "string",
					"example": ""
				},
				"DPT": {
					"type": "string",
					"example": ""
				},
				"VILLE_M": {
					"type": "string",
					"example": ""
				},
				"NOMOFF": {
					"type": "string",
					"example": ""
				},
				"LOCA": {
					"type": "string",
					"example": ""
				},
				"LOCA2": {
					"type": "string",
					"example": ""
				},
				"LOCA3": {
					"type": "string",
					"example": ""
				},
				"MILL": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"MILU": {
					"type": "string",
					"example": ""
				},
				"MOSA": {
					"type": "string",
					"example": ""
				},
				"MSGCOM": {
					"type": "string",
					"example": ""
				},
				"MUSEO": {
					"type": "string",
					"example": ""
				},
				"NSDA": {
					"type": "string",
					"example": ""
				},
				"ONOM": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PAUT": {
					"type": "string",
					"example": ""
				},
				"PDAT": {
					"type": "string",
					"example": ""
				},
				"PDEC": {
					"type": "string",
					"example": ""
				},
				"PEOC": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PERI": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PERU": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PHOT": {
					"type": "string",
					"example": ""
				},
				"PINS": {
					"type": "string",
					"example": ""
				},
				"PLIEUX": {
					"type": "string",
					"example": ""
				},
				"PREP": {
					"type": "string",
					"example": ""
				},
				"PUTI": {
					"type": "string",
					"example": ""
				},
				"RANG": {
					"type": "string",
					"example": ""
				},
				"REDA": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REFIM": {
					"type": "string",
					"example": ""
				},
				"REFMEM": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REFMER": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REFPAL": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REPR": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"RETIF": {
					"type": "string",
					"example": ""
				},
				"SREP": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"STAT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"TECH": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"TICO": {
					"type": "string",
					"example": ""
				},
				"TITR": {
					"type": "string",
					"example": ""
				},
				"TOUT": {
					"type": "string",
					"example": ""
				},
				"UTIL": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"VIDEO": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"WWW": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"LVID": {
					"type": "string",
					"example": ""
				},
				"_id": {
					"type": "string",
					"example": ""
				},
				"REF": {
					"type": "string",
					"example": ""
				},
				"__v": {
					"type": "number",
					"example": 0
				},
				"HISTORIQUE": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"_id": {
								"type": "string",
								"example": ""
							},
							"nom": {
								"type": "string",
								"example": ""
							},
							"prenom": {
								"type": "string",
								"example": ""
							},
							"email": {
								"type": "string",
								"example": ""
							},
							"date": {
								"type": "string",
								"example": ""
							},
							"updateMode": {
								"type": "string",
								"example": ""
							}
						}
					}
				}
			}
		},
...
	}
}